prototipo-simulado/requirements.txt
Flask==3.1.0
flask-cors==6.0.1
pandas==2.2.0
openpyxl==3.1.5
google-generativeai==0.8.3
gtts==2.5.4
Werkzeug==3.1.0
requests==2.31.0
gspread==6.1.0
google-auth==2.27.0
